requirements.txt
openai
azure-ai-inference
boto3
pypdf
pyserini
matplotlib
matplotlib-venn
pandas
upsetplot
sentence_transformers
seaborn
dotenv